MCPJVM-49: Deterministic Synthesis Refactor: Introduce Internal Framework Synthesizer Plugins Behind probe_recipe_create
