+++ java-agent/core/src/main/java/com/nimbly/mcpjvmdebugger/agent/AgentConfig.java
