+++ java-agent/pom.xml
@@ -5,55 +5,12 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.nimbly</groupId>
-  <artifactId>mcp-jvm-probe-agent</artifactId>
+  <artifactId>mcp-jvm-probe-agent-parent</artifactId>
   <version>0.1.0</version>
-  <packaging>jar</packaging>
-  <name>mcp-jvm-probe-agent</name>
+  <packaging>pom</packaging>
+  <name>mcp-jvm-probe-agent-parent</name>
 
-  <properties>
-    <maven.compiler.release>17</maven.compiler.release>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <bytebuddy.version>1.14.19</bytebuddy.version>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>net.bytebuddy</groupId>
-      <artifactId>byte-buddy</artifactId>
-      <version>${bytebuddy.version}</version>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>3.5.3</version>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <createDependencyReducedPom>false</createDependencyReducedPom>
-              <shadedArtifactAttached>true</shadedArtifactAttached>
-              <shadedClassifierName>all</shadedClassifierName>
-              <transformers>
-                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <manifestEntries>
-                    <Premain-Class>com.nimbly.mcpjvmdebugger.agent.ProbeAgent</Premain-Class>
-                    <Can-Redefine-Classes>true</Can-Redefine-Classes>
-                    <Can-Retransform-Classes>true</Can-Retransform-Classes>
-                  </manifestEntries>
-                </transformer>
-              </transformers>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
+  <modules>
+    <module>core</module>
+  </modules>
 </project>
-
