+++ src/models/synthesis/synthesizer_input.model.ts
@@ -1,0 +1,15 @@
+export type SynthesizerIntentMode =
+  | "regression_api_only"
+  | "single_line_probe"
+  | "regression_plus_line_probe";
+
+export type SynthesizerInput = {
+  rootAbs: string;
+  workspaceRootAbs: string;
+  searchRootsAbs: string[];
+  classHint: string;
+  methodHint: string;
+  intentMode: SynthesizerIntentMode;
+  lineHint?: number;
+  inferredTargetFileAbs?: string;
+};
