+++ src/tools/core/probe/handler.ts
