+++ src/tools/core/probe_check/contract.ts
@@ -1,4 +1,4 @@
-import { ProbeDiagnoseInputSchema } from "../../models/inputs";
+import { ProbeDiagnoseInputSchema } from "../../../models/inputs";
 
 export const PROBE_CHECK_TOOL = {
   name: "probe_check",
