+++ src/tools/core/probe_check/handler.ts
