+++ src/tools/synthesizers/registry/plugin.contract.ts
@@ -1,0 +1,13 @@
+import type { SynthesizerFailure } from "../../../models/synthesis/synthesizer_failure.model";
+import type { SynthesizerInput } from "../../../models/synthesis/synthesizer_input.model";
+import type { SynthesizerOutput } from "../../../models/synthesis/synthesizer_output.model";
+
+export type SynthesizerResult = SynthesizerOutput | SynthesizerFailure;
+
+export interface SynthesizerPlugin {
+  id: string;
+  framework: string;
+  pluginApiVersion: string;
+  canHandle(input: SynthesizerInput): Promise<boolean>;
+  synthesize(input: SynthesizerInput): Promise<SynthesizerResult>;
+}
