+++ src/tools/synthesizers/spring/adapter.ts
@@ -1,0 +1,1 @@
+export { springSynthesizerPlugin } from "./plugin";
