+++ src/utils/synthesizers/spring/failure_codes.util.ts
@@ -1,0 +1,4 @@
+export const SPRING_FAILURE_CODES = {
+  ENTRYPOINT_NOT_PROVEN: "spring_entrypoint_not_proven",
+  MAPPING_NOT_PROVEN: "spring_mapping_not_proven",
+} as const;
